refactor: Move common code to zepsec module
Create a zepsec module, and move the JIRA querying code into this
module.  This will allow different main programs to share this code.
